# EventManagementSystem.Client/OpenAPI/EventAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Event.Services",
    "version": "v1"
  },
  "paths": {
    "/api/Event": {
      "get": {
        "tags": [
          "Event"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Event/GetPublicEvents": {
      "get": {
        "tags": [
          "Event"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Event/search": {
      "get": {
        "tags": [
          "Event"
        ],
        "parameters": [
          {
            "name": "ID",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/Event/add": {
      "post": {
        "tags": [
          "Event"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EventModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EventModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/Event/upsert": {
      "put": {
        "tags": [
          "Event"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EventModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EventModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/Event/delete": {
      "delete": {
        "tags": [
          "Event"
        ],
        "parameters": [
          {
            "name": "ID",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Event/newID": {
      "get": {
        "tags": [
          "Event"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventTemporaryDataModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventTemporaryDataModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventTemporaryDataModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/EventComments": {
      "get": {
        "tags": [
          "EventComments"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventCommentsModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventCommentsModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventCommentsModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/EventComments/search": {
      "get": {
        "tags": [
          "EventComments"
        ],
        "parameters": [
          {
            "name": "ID",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventCommentsModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventCommentsModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventCommentsModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/EventComments/add": {
      "post": {
        "tags": [
          "EventComments"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCommentsModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCommentsModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EventCommentsModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventCommentsModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventCommentsModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventCommentsModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/EventComments/upsert": {
      "put": {
        "tags": [
          "EventComments"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCommentsModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCommentsModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EventCommentsModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventCommentsModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventCommentsModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventCommentsModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/EventComments/delete": {
      "delete": {
        "tags": [
          "EventComments"
        ],
        "parameters": [
          {
            "name": "ID",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/EventComments/newID": {
      "get": {
        "tags": [
          "EventComments"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventCommentsTemporaryDataModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventCommentsTemporaryDataModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventCommentsTemporaryDataModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/EventImage": {
      "get": {
        "tags": [
          "EventImage"
        ],
        "parameters": [
          {
            "name": "EventID",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventImageModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventImageModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventImageModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/EventImage/AddImage": {
      "post": {
        "tags": [
          "EventImage"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventImageModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EventImageModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EventImageModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/EventImage/AddImageRange": {
      "post": {
        "tags": [
          "EventImage"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EventImageModel"
                },
                "nullable": true
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EventImageModel"
                },
                "nullable": true
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EventImageModel"
                },
                "nullable": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/EventImage/deleteimage": {
      "delete": {
        "tags": [
          "EventImage"
        ],
        "parameters": [
          {
            "name": "ID",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "EventID",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EmployeeModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "fName": {
            "type": "string",
            "nullable": true
          },
          "lName": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "doB": {
            "type": "string",
            "format": "date-time"
          },
          "nic": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "address": {
            "type": "string",
            "format": "multiline",
            "nullable": true
          },
          "contact": {
            "type": "string",
            "format": "tel",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "empRole": {
            "type": "string",
            "nullable": true
          },
          "empPassWord": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EventModel": {
        "type": "object",
        "properties": {
          "eventID": {
            "type": "string",
            "nullable": true
          },
          "eventTitle": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "duration": {
            "type": "integer",
            "format": "int64"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "publisher": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "eventRecordHanledBy": {
            "$ref": "#/components/schemas/EmployeeModel"
          },
          "visibility": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EventTemporaryDataModel": {
        "type": "object",
        "properties": {
          "newID": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EventCommentsModel": {
        "required": [
          "comment",
          "commentedUser",
          "eventID",
          "id",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "eventID": {
            "$ref": "#/components/schemas/EventModel"
          },
          "commentedUser": {
            "$ref": "#/components/schemas/EmployeeModel"
          },
          "comment": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "EventCommentsTemporaryDataModel": {
        "type": "object",
        "properties": {
          "newID": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "EventImageModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "event": {
            "$ref": "#/components/schemas/EventModel"
          },
          "imageName": {
            "type": "string",
            "nullable": true
          },
          "isImageAvailable": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    }
  }
}