# EventManagementSystem.Data/Models/EmpTemporaryDataModel.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace EventManagementSystem.Models
{
    public class EmpTemporaryDataModel
    {
        public string NewID { get; set; }
    }
    public class EventTemporaryDataModel
    {
        public string NewID { get; set; }
    }
    public class EventCommentsTemporaryDataModel
    {
        public int NewID { get; set; }
    }

}